requirements.txt
python-dotenv

flask

sqlmodel
pydantic
email-validator

bcrypt

markdown2

httpx
pytest
ruff
